+++ NEWS.md
@@ -126,6 +126,10 @@
     overhead for both code and RAM, but it significantly simplifies the code.
   * Windows: The "xattr" tab now respects the LC_ALL and/or LC_MESSAGES
     environment variable to allow for easier multi-language testing.
+  * Windows: On ARM64 Windows 11, register for i386, arm64, and arm64ec.
+    * On ARM64 Windows 10, only register amd64 if using build 21277 or later.
+    * Fixes Fixes #398: Installing on ARM64 shows an error that the AMD64 version of the DLL couldn't be registered
+      * Reported by @kristibektashi.
 
 ## v2.2.1 (released 2023/07/30)